+++ wand/api.py
@@ -142,10 +142,12 @@ except (OSError, IOError):
     elif sys.platform == 'win32':
         msg += '#install-imagemagick-on-windows'
     elif sys.platform == 'darwin':
-        for pkgmgr in 'brew', 'port':
+        mac_pkgmgrs = {'brew': 'brew install freetype imagemagick',
+                       'port': 'port install imagemagick'}
+        for pkgmgr in mac_pkgmgrs:
             with os.popen('which ' + pkgmgr) as f:
                 if f.read().strip():
-                    msg = pkgmgr + ' install imagemagick'
+                    msg = mac_pkgmgrs[pkgmgr]
                     break
         else:
             msg += '#install-imagemagick-on-mac'